Miscelaneous fixes
Moved and completed some DMA/Hardware related functions, fixed a few warnings and made the startup routine actually clear all of WRAM because the C standard asks for uninitialised memory to be zero not random. I couldn't think of good way around hard coding the KDATA bank so I left it as is.